requirements.txt
Flask==2.2.2
speedtest-cli==2.1.3